Put some util methods into utilhandle to avoid circular import.
