+++ b2handle/tests/handleclient_writeaccess_patched_test.py
@@ -21,7 +21,7 @@ from b2handle.handleexceptions import ReverseLookupException
 from b2handle.handleexceptions import GenericHandleError
 from mockresponses import MockResponse, MockSearchResponse
 from utilities import failure_message, replace_timestamps, sort_lists
-from b2handle.util import check_handle_syntax
+from b2handle.utilhandle import check_handle_syntax
 
 class EUDATHandleClientWriteaccessPatchedTestCase(unittest.TestCase):
     '''Testing methods with write access (patched server access).
